requirements.txt
langchain
python-dotenv
pyodbc
azure-identity
langchain-openai
streamlit